+++ changelog.md
@@ -1,5 +1,9 @@
 # Mango Program Change Log
 
+## v3.2.15
+Deployed: | Slot:
+1. Impose price limits on spot orders
+
 ## v3.2.14
 Deployed: Jan 2, 2022 at 20:48:01 UTC | Slot: 114,518,931
 1. Check bids and asks when loading perp market book